# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MiniMarket Plus API",
    "description": "Backend para gestion de inventario, ventas, carrito y usuarios con seguridad JWT.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/carrito/{id}": {
      "get": {
        "tags": [
          "Carrito"
        ],
        "summary": "Obtiene un carrito por id con enlaces HATEOAS",
        "description": "Consulta un registro individual del carrito e incorpora enlaces hacia el recurso, la coleccion, el usuario y el producto.",
        "operationId": "obtenerCarritoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del registro de carrito",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "404": {
            "description": "Carrito no encontrado"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "200": {
            "description": "Carrito encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelCarritoDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Carrito"
        ],
        "summary": "Actualiza un carrito",
        "description": "Modifica la cantidad, usuario o producto asociado a un registro de carrito existente.",
        "operationId": "actualizarCarrito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del registro de carrito",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Carrito"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "404": {
            "description": "Carrito no encontrado"
          },
          "200": {
            "description": "Carrito actualizado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelCarritoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "400": {
            "description": "Solicitud invalida"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Carrito"
        ],
        "summary": "Elimina un producto del carrito",
        "description": "Elimina un registro del carrito cuando existe en la base de datos.",
        "operationId": "eliminarProductoDelCarrito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del registro de carrito",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "404": {
            "description": "Carrito no encontrado"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "204": {
            "description": "Producto eliminado del carrito correctamente"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/carrito": {
      "get": {
        "tags": [
          "Carrito"
        ],
        "summary": "Lista carritos con enlaces HATEOAS",
        "description": "Retorna registros de carrito como CollectionModel\u003CEntityModel\u003CCarritoDTO\u003E\u003E, incluyendo links self, collection, usuario y producto.",
        "operationId": "listarCarrito",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numero de pagina a consultar, comenzando desde 0",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Cantidad de registros por pagina",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 3
          }
        ],
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "200": {
            "description": "Carritos consultados correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelCarritoDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Carrito"
        ],
        "summary": "Agrega un producto al carrito",
        "description": "Registra un producto asociado a un usuario dentro del carrito de compras.",
        "operationId": "agregarProductoAlCarrito",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Carrito"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "200": {
            "description": "Producto agregado al carrito correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelCarritoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Rol": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rol"
            },
            "uniqueItems": true
          }
        }
      },
...
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "precio": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        }
      },
...
      "Carrito": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "EntityModelCarritoDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador unico del registro de carrito",
            "example": 1
          },
          "usuarioId": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador del usuario asociado al carrito",
            "example": 4
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario asociado al carrito",
            "example": "cliente"
          },
          "productoId": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador del producto agregado",
            "example": 1
          },
          "productoNombre": {
            "type": "string",
            "description": "Nombre del producto agregado",
            "example": "Arroz grano largo 1kg"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32",
            "description": "Cantidad del producto dentro del carrito",
            "example": 2
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "CollectionModelEntityModelCarritoDTO": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "carritos": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelCarritoDTO"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}